[PAPP-767] Moved the generic access tests to booth tests, since generic class is now abstract
